DeleteValidator

+++ ProductsManager/Application/Use Cases/CommandHandlers/DeleteProductCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using Application.Use_Cases.Commands;
 using Domain.Repositories;
+using FluentValidation;
 using MediatR;
 
 namespace Application.Use_Cases.CommandHandlers
@@ -15,6 +16,13 @@ namespace Application.Use_Cases.CommandHandlers
 
         public async Task<Guid> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
+            DeleteProductCommandValidator validatorRules = new DeleteProductCommandValidator();
+            var validator = validatorRules.Validate(request);
+            if (!validator.IsValid)
+            {
+                throw new ValidationException(validator.Errors);
+            }
+
             var product = await repository.GetProductAsync(request.Id);
             await repository.DeleteAsync(request.Id);
             return product.Id;
